Fix failed to build from source on Linux due to filename case insensitivity(windows.h)
Currently cross-building on Linux will fail during compilation stage
with the following error message:

```
[ 35%] Building CXX object CMakeFiles/wkbre.dir/cursor.cpp.obj
In file included from wkbre/anim.cpp:17:
wkbre/global.h:27:10: fatal error: Windows.h: No such file or directory
   27 | #include <Windows.h>
      |          ^~~~~~~~~~~
compilation terminated.
```

This is due to the filename casing of the "Windows.h" header file which
is not accessible in the case-sensitive file/operating systems like
GNU/Linux.

This patch fixes the problem by replacing them with "windows.h" as it
seems to be the proper filename casing in the Windows platform[1] as
well as in MinGW[2].

[1]: https://en.wikibooks.org/wiki/Windows_Programming/windows.h
[2]: https://stackoverflow.com/questions/1549123/how-to-use-the-windows-api-in-mingw

Signed-off-by: 林博仁(Buo-ren, Lin) <Buo.Ren.Lin@gmail.com>